create nav bar
