+++ src/App.vue
@@ -1,9 +1,11 @@
 <script setup lang="ts">
 import navbar from "./components/navbar.vue";
+import topbar from "./components/topbar.vue";
 </script>
 
 <template>
-  <div class="w-screen flex flex-col py-5 px-10">
+  <div class="w-screen flex flex-col">
+    <topbar />
     <navbar />
   </div>
 </template>
