+++ src/components/common/iconButton.vue
@@ -3,7 +3,7 @@ defineProps(["icon", "text"]);
 </script>
 
 <template>
-  <div class="flex gap-2 items-center">
+  <div class="flex gap-1 items-center">
     <img class="w-6 aspect-square" :src="icon" :alt="text" />
     <p>{{ text }}</p>
   </div>
